Added the implementation of accessing the server, depending on the need to sort the data on the page or not.
Combined the column sorting functions into one function.
Now sorting is carried out on the server side and occurs according to the principle of linking the values of data attributes (recorded in the corresponding table elements) with the variable "sort" (setState), the final values of which are used in requests to the server.

Added a hint to reset sorting.